requirements.txt
Flask
python-dotenv
openai
gunicorn
pypdf
